Changes before presentation
Aligned player books
Changed monster parameters
Fixed pickup controller
Made equipment respawn if dropped of the map
Added an invisible trigger collider for the mouse to track

+++ HellFryer/Assets/Scripts/Cooking/Orders/OrdersManager.cs
@@ -106,7 +106,7 @@ public class OrdersManager : MonoBehaviour
         {
             if (score != null)
             {
-                score.text = "Your score: " + totalScore;
+                score.text = "Your score: " + (int)totalScore;
                 score.gameObject.transform.parent.gameObject.SetActive(true);
             }
             //Change to next scene
